+++ src/Common/AppContext.cs
@@ -1,0 +1,6 @@
+﻿namespace Common;
+
+public class AppContextProxy : IAppContext
+{
+    public string BaseDirectory => AppContext.BaseDirectory;
+}