+++ src/Common/IAppContext.cs
@@ -1,0 +1,6 @@
+﻿namespace Common;
+
+public interface IAppContext
+{
+    string BaseDirectory { get; }
+}