+++ src/Recyclarr/Config/IConfigurationLoader.cs
@@ -5,7 +5,7 @@ namespace Recyclarr.Config;
 public interface IConfigurationLoader<out T>
     where T : IServiceConfiguration
 {
-    IEnumerable<T> Load(string propertyName, string configSection);
+    IEnumerable<T> Load(string file, string configSection);
     IEnumerable<T> LoadFromStream(TextReader stream, string configSection);
     IEnumerable<T> LoadMany(IEnumerable<string> configFiles, string configSection);
 }