+++ src/TrashLib/IConfigurationFinder.cs
@@ -1,0 +1,6 @@
+﻿namespace Recyclarr;
+
+public interface IConfigurationFinder
+{
+    string FindConfigPath();
+}